+++ CHANGELOG.md
@@ -27,6 +27,10 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * Navigation spans now use the span name `app.ui.navigation` (previously `screen name change`). #625
 
+### Fixed
+
+* Fixed xcframework build system. #647
+
 ## [2.2.3] - 2026-05-12
 
 ### Fixed